Adding the projection plane distance.

+++ src/math_utils.c
@@ -9,7 +9,7 @@ double deg_to_radian (int deg)
     return (deg * (M_PI / 180));
 }
 
-double radian_to_deg(double rad)
+int radian_to_deg(double rad)
 {
     return ((rad / M_PI) * 180);
 }
